+++ packages/cli/src/index.ts
@@ -1,4 +1,4 @@
-import { findAndParseConfig, getDefaultSyncImport } from '@graphql-mesh/config';
+import { findAndParseConfig } from '@graphql-mesh/config';
 import { getMesh, GetMeshOptions } from '@graphql-mesh/runtime';
 import { generateTsArtifacts } from './commands/ts-artifacts';
 import { serveMesh } from './commands/serve/serve';
@@ -6,7 +6,7 @@ import { isAbsolute, resolve, join } from 'path';
 import { existsSync } from 'fs';
 import { FsStoreStorageAdapter, MeshStore } from '@graphql-mesh/store';
 import { printSchemaWithDirectives } from '@graphql-tools/utils';
-import { writeFile, pathExists, rmdirs, DefaultLogger } from '@graphql-mesh/utils';
+import { writeFile, pathExists, rmdirs, DefaultLogger, getDefaultSyncImport } from '@graphql-mesh/utils';
 import { handleFatalError } from './handleFatalError';
 import { cwd, env } from 'process';
 import yargs from 'yargs';
